Remove BaseFoldApp again in favor of utitlity module
