+++ src/biotite/application/viennarna/util.py
@@ -1,0 +1,15 @@
+# This source code is part of the Biotite package and is distributed
+# under the 3-Clause BSD License. Please see 'LICENSE.rst' for further
+# information.
+
+__name__ = "biotite.application"
+__author__ = "Patrick Kunzmann"
+__all__ = ["build_constraint_string"]
+
+
+def build_constraint_string(self, sequence_length,
+                            pairs=None, paired=None, unpaired=None,
+                            downstream=None, upstream=None):
+    """
+    """
+    pass
